refactor: Make nextjs pages look more like controllers
