+++ frontend/@types/index.d.ts
@@ -2,3 +2,10 @@ declare module '*.svg' {
    const value: any;
    export default value;
 }
+
+declare type NextPageWithLayout<P = {}, IP = P> = import('next').NextPage<
+   P,
+   IP
+> & {
+   getLayout?: (page: React.ReactElement, pageProps: P) => React.ReactNode;
+};
