--- frontend/components/ProductList/index.ts
@@ -1,3 +1,0 @@
-export * from './PageLayout';
-export * from './MetaTags';
-export * from './sections';
