+++ frontend/components/common/Layout/SearchBar.tsx
