+++ frontend/components/product-list/sections/BannerSection/index.tsx
