+++ frontend/components/product-list/sections/FilterableProductsSection/AppliedFilters.tsx
