+++ frontend/components/product-list/sections/FilterableProductsSection/FilterList/FilterRow.tsx
