+++ frontend/components/product-list/sections/FilterableProductsSection/FilterList/RangeFilter/index.ts
