+++ frontend/components/product-list/sections/FilterableProductsSection/FilterList/virtualAccordion.machine.ts
