+++ frontend/components/product-list/sections/FilterableProductsSection/SearchInput.tsx
