+++ frontend/components/product-list/sections/HeroSection.tsx
