+++ frontend/components/product-list/sections/NewsletterSection/index.tsx
