+++ frontend/components/product-list/sections/ProductListChildrenSection.tsx
