+++ frontend/components/product-list/sections/RelatedPostsSection/index.tsx
