+++ frontend/components/product-list/sections/index.ts
