+++ frontend/components/ui/index.ts
@@ -1,0 +1,2 @@
+export * from './Card';
+export * from './Rating';
