+++ frontend/pages/store/[handle].tsx
@@ -1,25 +1,12 @@
-import { Layout } from '@components/Layout';
+import { Layout } from '@components/common';
 import {
-   BannerSection,
-   FeaturedProductListSection,
-   ProductListSetSection,
-   FilterableProductsSection,
-   HeroSection,
-   MetaTags,
-   NewsletterSection,
-   PageLayout,
-   RelatedPostsSection,
-   ProductListChildrenSection,
-} from '@components/ProductList';
-import { ALGOLIA_API_KEY, ALGOLIA_APP_ID } from '@config/env';
-import { AlgoliaProvider, SearchContext } from '@lib/algolia';
-import { getGlobalSettings, GlobalSettings } from '@models/global-settings';
+   ProductListView,
+   ProductListViewProps,
+} from '@components/product-list';
+import { getGlobalSettings } from '@models/global-settings';
 import {
    createProductListSearchContext,
    getProductListByHandle,
-   ProductSearchHit,
-   ProductList,
-   ProductListSectionType,
 } from '@models/product-list';
 import {
    getStoreByCode,
@@ -30,136 +17,16 @@ import {
 import { GetServerSideProps } from 'next';
 import * as React from 'react';
 
-interface ProductListPageProps {
-   productList: ProductList;
-   searchContext: SearchContext<ProductSearchHit>;
-   globalSettings: GlobalSettings;
+type PageProps = ProductListViewProps & {
    stores: StoreListItem[];
    currentStore: Store;
-}
-
-const ALGOLIA_DEFAULT_INDEX_NAME = 'shopify_ifixit_test_products';
-
-export default function ProductListPage({
-   productList,
-   searchContext,
-   globalSettings,
-}: ProductListPageProps) {
-   const { newsletterForm } = globalSettings;
-   return (
-      <PageLayout>
-         <AlgoliaProvider
-            key={productList.handle}
-            appId={ALGOLIA_APP_ID}
-            apiKey={ALGOLIA_API_KEY}
-            initialIndexName={ALGOLIA_DEFAULT_INDEX_NAME}
-            initialContext={searchContext}
-         >
-            <MetaTags productList={productList} />
-            <HeroSection productList={productList} />
-            {productList.children.length > 0 && (
-               <ProductListChildrenSection
-                  heading={productList.title}
-                  productListChildren={productList.children}
-               />
-            )}
-            <FilterableProductsSection />
-            {productList.sections.map((section, index) => {
-               switch (section.type) {
-                  case ProductListSectionType.Banner: {
-                     return (
-                        <BannerSection
-                           key={index}
-                           title={section.title}
-                           description={section.description}
-                           callToActionLabel={section.callToActionLabel}
-                           url={section.url}
-                        />
-                     );
-                  }
-                  case ProductListSectionType.RelatedPosts: {
-                     const tags = [productList.title].concat(
-                        section.tags?.split(',').map((tag) => tag.trim()) || []
-                     );
-                     return <RelatedPostsSection key={index} tags={tags} />;
-                  }
-                  case ProductListSectionType.FeaturedProductList: {
-                     const { productList } = section;
-                     if (productList) {
-                        return (
-                           <FeaturedProductListSection
-                              key={index}
-                              handle={productList.handle}
-                              algoliaIndexName={ALGOLIA_DEFAULT_INDEX_NAME}
-                              title={productList.title}
-                              description={productList.description}
-                              imageSrc={
-                                 productList.image
-                                    ? `${productList.image.url}`
-                                    : undefined
-                              }
-                              imageAlt={
-                                 productList.image?.alternativeText || undefined
-                              }
-                           />
-                        );
-                     }
-                     return null;
-                  }
-                  case ProductListSectionType.ProductListSet: {
-                     const { title, productLists } = section;
-                     if (productLists.length > 0) {
-                        return (
-                           <ProductListSetSection
-                              key={index}
-                              title={title}
-                              productLists={productLists}
-                           />
-                        );
-                     }
-                     return null;
-                  }
-                  default: {
-                     console.warn(
-                        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                        // @ts-ignore
-                        `Section ${section.__typename} not implemented`
-                     );
-                     return null;
-                  }
-               }
-            })}
-            {newsletterForm && (
-               <NewsletterSection
-                  title={newsletterForm.title}
-                  description={newsletterForm.subtitle}
-                  emailPlaceholder={
-                     newsletterForm.inputPlaceholder || undefined
-                  }
-                  subscribeLabel={newsletterForm.callToActionButtonTitle}
-               />
-            )}
-         </AlgoliaProvider>
-      </PageLayout>
-   );
-}
-
-ProductListPage.getLayout = function getLayout(
-   page: React.ReactElement,
-   pageProps: ProductListPageProps
-) {
-   return (
-      <Layout
-         title={`iFixit | ${pageProps.productList.title}`}
-         currentStore={pageProps.currentStore}
-         stores={pageProps.stores}
-      >
-         {page}
-      </Layout>
-   );
 };
 
-export const getServerSideProps: GetServerSideProps<ProductListPageProps> = async (
+// This constant should probably be a field of the store model (editable from CMS)
+// so that it's configurable per-store.
+const ALGOLIA_DEFAULT_INDEX_NAME = 'shopify_ifixit_test_products';
+
+export const getServerSideProps: GetServerSideProps<PageProps> = async (
    context
 ) => {
    // The data is considered fresh for 10 seconds, and can be served even if stale for up to 10 minutes
@@ -216,3 +83,31 @@ export const getServerSideProps: GetServerSideProps<ProductListPageProps> = asyn
       },
    };
 };
+
+const ProductListPage: NextPageWithLayout<PageProps> = ({
+   productList,
+   searchContext,
+   globalSettings,
+}) => {
+   return (
+      <ProductListView
+         productList={productList}
+         searchContext={searchContext}
+         globalSettings={globalSettings}
+      />
+   );
+};
+
+ProductListPage.getLayout = function getLayout(page, pageProps) {
+   return (
+      <Layout
+         title={`iFixit | ${pageProps.productList.title}`}
+         currentStore={pageProps.currentStore}
+         stores={pageProps.stores}
+      >
+         {page}
+      </Layout>
+   );
+};
+
+export default ProductListPage;
